# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIS LARAVEL-MALVA",
        "description": "APIS LARAVEL-MALVA",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/v1/auth/create": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Create an account",
                "operationId": "5290df7a649cb3077fd9ec32111513f2",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Idioma",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "dni",
                                    "email",
                                    "password",
                                    "id_rol"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Daniel"
                                    },
                                    "dni": {
                                        "type": "number",
                                        "example": "74112299"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "aldanagerardo24@gmail.com"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "https://res.cloudinary.com/dly4rnmgh/image/upload/v1730668603/imagenes/ckboys2cdnapznjb0zsl.png"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "id_rol": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Se registro correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto por un usuario existente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El usuario ya existe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Se supero el limite de peticiones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Se supero el limite de peticiones"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error Interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ocurrio un error,intentelo nuevamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "98f0c6061f7ddf2651d66f6ee1082856",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Idioma",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "aldanagerardo24@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "1|a54fds45fa4554f45adgssGADG5S4GSF54GS5FG4SFGSF"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cuerpo de peticion incorrecta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El usuario no existe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Las credenciales son incorrectas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Rechazo de solicitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El usuario no se encuentra activado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Se supero el limite de peticiones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Se supero el limite de peticiones"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error Interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ocurrio un error,intentelo nuevamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Return the user information",
                "operationId": "d2b1d6390b2ac117a6b5a9f4b8bf183a",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Idioma",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "dni": {
                                            "type": "number",
                                            "example": "12345678"
                                        },
                                        "id_rol": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Daniel"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "aldanagerardo24@gmail.com"
                                        },
                                        "estado": {
                                            "type": "number",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Las credenciales son incorrectas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Se superó el límite de peticiones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Se superó el límite de peticiones"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error Interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ocurrió un error, inténtelo nuevamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token to access these API endpoints",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}